corrected muddled logic for tax loss carryover, fixes #64
The tax losses used or lost are the greater of losses due to the
seven-year limit and the taxes that need to be deferred.

We use an exponential approximation for the losses due to the seven-year
limit.  Adding the MARCS depreciation tables (or even doing strictly
linear depreciation) would involve adding one stock for each year when
taxes would be carried over. The carryover would have to be drawn down
using compicated logic that draws on the oldest first. Adding such
complex logic for a minor part of the model isn't warranted.